+++ commons/dimas-com/src/lib.rs
@@ -4,8 +4,8 @@
 //!
 
 // region:		--- modules
-use crate::error::{DimasError, Result};
 use bitcode::{encode, Encode};
+use dimas_core::error::{DimasError, Result};
 use std::fmt::Debug;
 use std::sync::Arc;
 use zenoh::prelude::{r#async::*, sync::SyncResolve};
@@ -26,7 +26,7 @@ impl Communicator {
 	/// Constructor
 	/// # Errors
 	///
-	pub fn new(config: crate::config::Config) -> Result<Self> {
+	pub fn new(config: dimas_config::Config) -> Result<Self> {
 		let cfg = config;
 		let session = Arc::new(
 			zenoh::open(cfg.zenoh_config())
@@ -106,6 +106,20 @@ impl Communicator {
 			.res_sync()
 			.map_err(|_| DimasError::Delete.into())
 	}
+
+	/// Send an ad hoc query using the given `topic`.
+	/// The `topic` will be enhanced with the group prefix.
+	/// Answers are collected via callback
+	/// # Errors
+	///
+	pub fn get(&self, topic: &str) -> Result<()> {
+		let key_expr = self.key_expr(topic);
+
+		self.session
+			.delete(&key_expr)
+			.res_sync()
+			.map_err(|_| DimasError::Delete.into())
+	}
 }
 // endregion:	--- Communicator
 
@@ -125,7 +139,7 @@ mod tests {
 	#[tokio::test(flavor = "multi_thread")]
 	//#[serial]
 	async fn communicator_create_multi() -> Result<()> {
-		let mut peer1 = Communicator::new(crate::config::Config::default())?;
+		let mut peer1 = Communicator::new(dimas_config::Config::default())?;
 		peer1.set_prefix("peer1");
 		Ok(())
 	}