+++ dimas/src/lib.rs
@@ -5,7 +5,6 @@
 #![warn(missing_docs)]
 #![doc = include_str!("../README.md")]
 
-//#[cfg(feature = "nightly")]
 #[cfg(doctest)]
 doc_comment::doctest!("../README.md");